demo

+++ registroCursos/views.py
@@ -6,7 +6,6 @@ from django.contrib.auth.models import User
 from .forms import *
 
 
-
 date_filter = timezone.now().strftime('%d/%m/%Y')
 currentSemester = "ago_dic" if int(date_filter[-7:-5]) > 7 else "ene_jun"
 
@@ -16,9 +15,9 @@ class UserProfileView(View):
 
     def post(self, request, *args, **kwargs):
         post_data = request.POST.copy()
-        
+
         user_profile = UserProfile.objects.get(id=request.user.id)
-        
+
         user = User.objects.get(id=request.user.id)
         print(request.user.id)
         if User.objects.filter(username=post_data['username']) and not User.objects.filter(id=request.user.id, username=post_data['username']):
@@ -40,9 +39,9 @@ class UserProfileView(View):
         return render(request, 'user/update_user_profile_done.html', locals())
 
     def get(self, request, *args, **kwargs):
-        
+
         profile = UserProfile.objects.get(id=request.user.id)
-       
+
         form_profile = UpdateUserProfileForm()
         form_user = UpdateUserForm()
 
@@ -93,39 +92,67 @@ class DetailView(View):
     template_name = "detail.html"
 
     def post(self, request, *args, **kwargs):
+
         return render(request, self.template_name, locals())
 
     def get(self, request, *args, **kwargs):
-        current_user = request.user
-        current_user_id = current_user.id
+
         id = self.kwargs.get('id_curso')
         curso = Curso.objects.get(id=id)
         date = date_filter
+
         onCourseAsInstructor = True if Curso.objects.filter(
-            id=id, instructor=current_user_id) else False
+            id=id, instructor=request.user.id) else False
+
         onCourseAsAlumno = True if Curso.objects.filter(
-            id=id, alumno=current_user_id) else False
+            id=id, alumno=request.user.id) else False
+
         onCourseAsColaborador = True if Curso.objects.filter(
-            id=id, colaborador=current_user_id) else False
+            id=id, colaborador=request.user.id) else False
+
         return render(request, self.template_name, locals())
 
 
-class DisEnrollCourseView(View):
-    template_name = 'user/confirm.html'
+class ConfirmActionCourseView(View):
+    template_name = 'user/course_confirm_action.html'
 
     def post(self, request, *args, **kwargs):
 
         post_data = request.POST.copy()
+
         curso = Curso.objects.get(id=post_data['curso'])
         usuario = User.objects.get(id=post_data['user'])
 
-        if Historial.objects.filter(curso=curso, alumno=usuario):
-            Historial.objects.get(curso=curso.id, alumno=usuario.id).delete()
+         # QuerySet que verifica que no tenga otro grupo inscrito a la misma hora
+        verify_enroll = Curso.objects.filter(alumno=request.user.id,
+                                            hora_inicio=curso.hora_inicio,
+                                            dia_inicio=curso.dia_inicio).exclude(id=curso.id)
+                                            
+        # variable que confirma la accion
+        done = False
+
+            # Si el usuario desea inscribirse a un curso
+        if post_data['action'] == 'enroll':
+            # QuerySet que verifica que no este inscrito ya 
+            if not Historial.objects.filter(curso=curso,
+             alumno=usuario):
+                if not verify_enroll: 
+                    h1 = Historial(curso=curso,
+                               alumno=usuario)
+                    h1.save()
+                    done = True
+
+        if post_data['action'] == 'disenroll':
+            if Historial.objects.filter(curso=curso, alumno=usuario):
+                Historial.objects.get(curso=curso.id, alumno=usuario.id).delete()
+                done = True
 
         return render(request, self.template_name, locals())
+
 
     def get(self, request, *args, **kwargs):
         return render(request, self.template_name, locals())
+
 
 
 class CursosView(View):
@@ -144,26 +171,6 @@ class CursosView(View):
         return render(request, self.template_name, locals())
 
 
-class EnrollCourseView(View):
-
-    template_name = 'user/confirm.html'
-
-    def post(self, request, *args, **kwargs):
-        post_data = request.POST.copy()
-        curso = Curso.objects.get(id=post_data['curso'])
-        usuario = User.objects.get(id=post_data['user'])
-
-        if not Historial.objects.filter(curso=curso, alumno=usuario):
-            h1 = Historial(curso=curso, alumno=usuario)
-            h1.save()
-
-        return render(request, self.template_name, locals())
-
-    def get(self, request, *args, **kwargs):
-
-        return render(request, self.template_name, locals())
-
-
 class HomeView(View):
 
     template_name = "home.html"
@@ -173,17 +180,14 @@ class HomeView(View):
 
     def get(self, request, *args, **kwargs):
 
-        current_user = request.user
-        current_user_id = current_user.id
-
-        cursos_alumno = Curso.objects.filter(alumno=current_user_id,
+        cursos_alumno = Curso.objects.filter(alumno=request.user.id,
                                              anno=int(date_filter[-4:]),
                                              semestre=currentSemester)
-        cursos_instructor = Curso.objects.filter(instructor_id=current_user_id,
+        cursos_instructor = Curso.objects.filter(instructor_id=request.user.id,
                                                  anno=int(date_filter[-4:]),
                                                  semestre=currentSemester)
 
-        cursos_colaborador = Curso.objects.filter(colaborador_id=current_user_id,
+        cursos_colaborador = Curso.objects.filter(colaborador_id=request.user.id,
                                                   anno=int(date_filter[-4:]),
                                                   semestre=currentSemester)
         return render(request, self.template_name, locals())
